+++ internal/converter/converter_interfaces.go
@@ -1,6 +1,6 @@
 package converter
 
-import "github.com/firstBitSportivnaya/files-converter/pkg/config"
+import "github.com/firstBitSportivnaya/files-converter/internal/config"
 
 type Converter interface {
 	Convert(cfg *config.Configuration) error
